backend-python/requirements.txt
kafka-python==2.1.4
geopy===2.3.0